refactor: encapsulate rtk logic and add curl fallback
